Move toward "current-xyz" keys for tracking what's happening "now"
- this makes it easier to access one thing at a time
- we only need to worry about the immediate point in the story
- as per bronek's suggestion - we great a "pointer" to where we are
via "current-step" an index of what step we are at in the current room.

TODO! currently the "current" keys ("current-room", "current-step") are
holding the entire data structure (the story text, commands, potential media
assets etc) -- we will break this out into a subscription JUST for displaying
the current-step in the story, and a sub for the ds itself.